requirements.txt
attrs==20.3.0
autobahn==21.1.1
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
constantly==15.1.0
cryptography==3.3.1
dateparser==1.0.0
hyperlink==21.0.0
idna==2.10
incremental==17.5.0
joblib==1.0.0
KuCoin==0.1
numpy==1.20.0
pandas==1.2.4
patsy==0.5.1
py-postgresql==1.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==20.0.1
Pypubsub==4.0.3
python-binance==0.7.9
python-dateutil==2.8.1
python-kucoin==2.1.2
pytz==2021.1
redis==3.5.3
regex==2020.11.13
requests==2.25.1
rq==1.7.0
scikit-learn==0.24.1
scipy==1.6.0
service-identity==18.1.0
six==1.15.0
statsmodels==0.12.2
threadpoolctl==2.1.0
Twisted==20.3.0
txaio==20.12.1
tzlocal==2.1
ujson==4.0.2
urllib3==1.26.3
websockets==9.0.2
zope.interface==5.2.0
